+++ CHANGES.rst
@@ -10,6 +10,13 @@ Unreleased
     switching ``POST`` to ``GET``. This preserves the current behavior of
     ``GET`` and ``POST`` redirects, and is also correct for frontend libraries
     such as HTMX. :pr:`3092`
+-   The test client clears more request body information when a redirect
+    switches to ``GET``. :pr:`3092`
+-   The test client only switches ``301`` and ``302`` redirects to ``GET`` if
+    the request was ``POST``. :pr:`3092`
+-   The test client does not handle ``305`` as a redirect, as it is no longer
+    part of the HTTP spec. :pr:`3092`
+
 
 Version 3.1.5
 -------------